Safari-609.1.11.1/LayoutTests/fast/dom/Attr/value-not-nullable-expected.txt

Tests that Attr.value is not nullable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


attr = document.createAttribute('test')
PASS attr.value is ""
PASS attr.value = null did not throw exception.
PASS attr.value is "null"
PASS attr.value = undefined did not throw exception.
PASS attr.value is "undefined"
PASS successfullyParsed is true

TEST COMPLETE
